image-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
py-eureka-client==1.2.6
python-dotenv==1.0.1
requests==2.31.0
